Refacor LoginForm into its own component
